Add test page for javascript clients
I'm adding a simple html test page and an older version of jQuery. This
is for testing and documenting how a front-end client can use these
endpoints.

The version of jQuery is the same as what's currently on the Classroom
site, because we know they will be a client of this service. And the
test html page, unsurprisingly, mirrors what we need to implement on the
Classroom page.