Adds the term and term_source models
The Term module uses TermSource to get a list of all defined terms.
TermSouce executes a SQL query to get all sessions defined in the
asr warehouse.  Instead of inheriting from ::ActiveRecord::Base, which expects a
term_source table to exist, we are rewriting some of the ActiveRecord
methods so that we can use the find_by_sql method. The find_by_sql
method expects a columns_hash, a connection, a table_name, and an
instantiate method as of Rails 4.2 - this could change in Rails 5. We
are only calling sanitize_sql because without that call find_by_sql
complains.