Update to ensure only unique term objects are created
The SQL query for TermSource returns a row for each session that is
defined, which means the term is duplicated across the rows. We were
creating a term object for every row returned from the SQL query, which
is not what we want.  By defining an == method to the Term model, we can
call the to_set method on the result set returned by find_by_sql so that
we only create terms for unique strm attributes.