Change rabl templates to conform with jsonapi
The expected return for for all of the routes is

```json
{[plural resource name]:[{resource},...]
```

as per jsonapi. The rabl templates were leaving out the plural resource
name at the start and not wrapping the whole thing in a {}.

The result of this is that jquery couldn't get the collection of
resources out correctly with syntax like `response.sessions`.

The small tweaks to the rabl templates are documented here:

https://github.com/nesquena/rabl/wiki/Conforming-to-jsonapi.org-format#resource-collections