Sort sessions by session_code
Fixes #19

The sort was pretty straightforward. Testing is a bit odd. The main
(only?) client of this code only cares about sessions in the context of
a term/institution/career. So the test ensures that results in that
context are ordered correctly.

But, more generally, this change means that if you go to `sessions.json`
without any filters, the results will be sorted by the session code,
which might be a little strange. I don't think it's a big problem, since
the usage of this service is so small.