Restore join that restricts terms returned
I took this join out because we never used any of the data on the joined
table. But the join was there to restrict the number of terms returned.
Without it the service returns terms going back to Spring, 1900.
Restoring the join and adding a comment to explain why it's there