Update Readme with new structure
In order to update the sessions API to the JSON standard according to
jsonapi.org, we are proposing changing the structure of the JSON
returned. We think this brings it into compliance. There are nested
objects we are removing, but if there is a reason we should keep them we
can. We think removing them would simplify parsing the response.

Co-authored-by: Brittany Greaner <grea0022@umn.edu>
Co-authored-by: Sierra Schultz <schu3450@umn.edu>
Co-authored-by: Tony Thomas <thoma127@umn.edu>