Update rack-cors and other tooling
**Locked `rack-cors` to version 1.0.6**

When we deployed to staging, the server gave us this error: `Incomplete
response received from application`. Running the tests, we noticed the
tests were giving us the following errors: `undefined method
'unescape_path' for Rack::Utils:Module`.

[We found this issue](https://github.com/cyu/rack-cors/issues/191) which said that
1.0.6 version fixes these errors. After locking to 1.0.6 all the tests
pass.

**Added Oracle Cleaner**

We noticed there were TFMS database tables in our development and test
databases. To eliminate this as a possibility for troubleshooting, we
introduced Oracle Cleaner to the project.

**Updated `db/schema.rb` with boilerplate**

This will keep showing as a change in Git if we don't commit it as is.
As it's just boilerplate, we thought it would be safe to add.

Co-authored-by: Brittany Greaner <grea0022@umn.edu>
Co-authored-by: Tony Thomas <thoma127@umn.edu>