Request spec updates
* Set up before block so we only make the request once
* Update tests to compare arrays instead of an array to a set

After switching to Rails 2.5, tests that compare a set to an array using
`match` started to fail. The purpose of converting the array to a set is
to limit the array to unique values. We can then convert the result back
to an array so we're comparing the same type of object.