Rails 7.1 Upgrade
Mostly text and syntax changes.

There was a deprecation with secrets (used for storing the secret_key_base)
in favor of credentials. We found that there was some Rails magic to this,
as renaming the config/secrets.yml file to config/credentials.yml fixed
the error.

With the upgrade to activerecord-oracle-adapter, the database.yml files
must be updated to remove _enhanced.

Co-authored-by: Cassie Simpson <simps614@umn.edu>