Remove commit SHA from script/prod_build
This PR modifies `script/prod_build` so that it only creates the `latest` tag when the script is run. Originally, we thought it would be a good idea to include an image tag that denotes the Git commit SHA in addition to the `latest` version, but this is not necessary because Kamal already automatically does this. Also, we already include the Git commit SHA in the Docker image via the `GIT_COMMIT` environment variable. This change eliminates the extra tagged images that get generated locally when the `script/prod_build` helper is executed so that there are fewer Docker images that need to be cleaned up over time.