Create `script/prod_run`
This PR implements [ClickUp task 86b6609c9](https://app.clickup.com/t/86b6609c9). It creates a `script/prod_run` file that can be used to run the staging/production flavor of the Docker image locally.

The following script was added to help facilitate the use of Vault:

* `script/_vault` - A helper that adds aliases to `vault` and `vault_saml_login` and uses the official HashiCorp Vault enterprise Docker image to run the Vault CLI

To test this new helper, perform the following steps:

1) `script/prod_run staging`
1) Open a web browser and navigate to: `http://localhost:3000/sessions.json`

NOTE: Commenting out `config.force_ssl = true` in the `config/environments/production.rb` was necessary because Puma will attempt to redirect you to an HTTPS endpoint otherwise. For local testing, this does not work because there is no reverse proxy to serve the site over SSL, and the web browser will try to send you to a SSL-enabled URL that doesn't exist. In practice, this is a safe change because Traefik already has rules in place to auto-redirect HTTP requests to HTTPS.