Update script/setup to use HCP Vault
This PR implements [ClickUp task 86b667drg](https://app.clickup.com/t/86b667drg). It removes the Lastpassify gem in favor of using HCP Vault to populate the development secrets.

The `config/database.development.yml.erb` file is an ERB template that is populated by calling `script/_setup` with a valid HCP Vault token. The `script/setup` helper was modified to prompt the user to login to HCP Vault at the very beginning of the script. The call to `docker compose run --rm web script/_setup $(whoami)` was changed to run as `root` and mount the `vault_cli` volume into the container so that `script/_setup` has both permissions and access to the `.vault-token` file that is created by the `vault_saml_login` alias.